client_requirements.txt
# Local dependencies for the StreamSync Live client
google-genai>=0.8.0
mss>=9.0.1
Pillow>=10.3.0
pyaudio>=0.2.14
